# docs/api/openapi-spec/swagger.yaml
definitions:
  dtos.GetPostResponseDto:
    description: GetPostResponseDto is a data transfer object for reporting the details
      of a post
    properties:
      post:
        $ref: '#/definitions/dtos.Post'
    type: object
  dtos.Post:
    properties:
      content:
        description: "@Description\tContent is the content of the post"
        type: string
      description:
        description: "@Description\tDescription is the description of the post"
        type: string
      status:
        description: "@Description\tStatus is the status of the post"
        type: string
      title:
        description: "@Description\tTitle is the title of the post"
        type: string
    type: object
  errors.appError:
    properties:
      cause:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
      status:
        type: integer
      type:
        type: string
    type: object
...
  http.CreatePostRequestDto:
    description: CreatePostRequestDto is a data transfer object for creating a post
    properties:
      content:
        description: "@Description\tContent is the content of the post"
        minLength: 10
        type: string
      description:
        description: "@Description\tDescription is the description of the post"
        minLength: 10
        type: string
      title:
        description: "@Description\tTitle is the title of the post"
        minLength: 10
        type: string
      user_id:
        description: "@Description\tUserId is the ID of the user creating the post"
        type: string
    required:
    - content
    - description
    - title
    - user_id
    type: object
  http.CreatePostResponseDto:
    description: CreatePostResponseDto is a data transfer object for reporting the
      details of a created post
    properties:
      content:
        description: "@Description\tContent is the content of the post"
        type: string
      description:
        description: "@Description\tDescription is the description of the post"
        type: string
      status:
        description: "@Description\tStatus is the status of the post"
        type: string
      title:
        description: "@Description\tTitle is the title of the post"
        type: string
    type: object
  http.GetPostResponseDto:
    description: GetPostResponseDto is a data transfer object for reporting the details
      of a post
    properties:
      posts:
        items:
          $ref: '#/definitions/http.Post'
        type: array
    type: object
...
  http.Post:
    properties:
      content:
        description: "@Description\tContent is the content of the post"
        type: string
      description:
        description: "@Description\tDescription is the description of the post"
        type: string
      status:
        description: "@Description\tStatus is the status of the post"
        type: string
      title:
        description: "@Description\tTitle is the title of the post"
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: It is a method to retrieve a pagination keys such as page and page_size
        and list posts
      parameters:
      - description: page number
        in: path
        name: page
        required: true
        type: string
      - description: page size
        in: path
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post listed successfully
          schema:
            $ref: '#/definitions/http.GetPostResponseDto'
      summary: List posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: It is a method to create a new post
      parameters:
      - description: Create Post Request DTO
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/http.CreatePostRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Post created successfully
          schema:
            $ref: '#/definitions/http.CreatePostResponseDto'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/errors.appError'
        "422":
          description: Post validation request body does not validate
          schema:
            $ref: '#/definitions/errors.appError'
        "500":
          description: Post create failed
          schema:
            $ref: '#/definitions/errors.appError'
      summary: post store
      tags:
      - posts
  /posts/{id}:
    get:
      consumes:
      - application/json
      description: It is a method to retrieve a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post retrieved successfully
          schema:
            $ref: '#/definitions/dtos.GetPostResponseDto'
        "400":
          description: The Id does not parse correctly
          schema:
            $ref: '#/definitions/errors.appError'
        "404":
          description: Page not found
          schema:
            $ref: '#/definitions/errors.appError'
      summary: Get post by id
      tags:
      - posts
...
swagger: "2.0"